Simplify the Source docs homepage
We remove installation instructions from the homepage as they add very
little and are repeated on the Foundations and Components pages.

We also remove some of the technical information describing the project,
as this is not the first thing a consumer of these packages would need
to know.

Without any of this detail, the home page is no longer a "Getting
Started" page, although it still serves as a landing page. Therefore
we rename it to "Source" to reflect its primacy in the information
architecture.