Use storybook-deployer
We are currently using the not-very-frequently-updated maxheld83/ghpages
Action to deploy Storybook to Github Pages. Recently this action started
failing for unknown reasons, so we thought we'd take the opportunity to
explore alternatives.

storybook-deployer is recommended and maintained by Storybook, so is
likely a good migration target for this task.