Visually hide `figcaption` in StarRating (#1624)
* Visually hide StarRating figcaption

The figcaption was being visually hidden from the DOM via a -1000px text
indent. This was adopted when the `<figure>` had an explicit height set,
with the stars being added as a background, and no other block or inline
content in the `<figure>`. Now that we're using `<svg>` for the stars,
the `<figcaption>` stacks below it, adding height to the component.

`visuallyHidden` styles remove the element's height, while accomplishing
the same effect as the previous -1000px text indent.

* Use plural in caption when there are zero stars

'zero stars', 'one star', '(n > 1) stars'

* Remove aria-role=complementary

* Add changesets

* Changeset 'minor'->'major'

Co-authored-by: TJ Silver <15648334+tjsilver@users.noreply.github.com>